rename udi_hid_generic to udi_hid_hww
