+++ src/CMakeLists.txt
@@ -87,7 +87,7 @@ set(DRIVER-SOURCES
         drivers/common/services/storage/ctrl_access/ctrl_access.c 
         drivers/common/services/usb/class/hid/device/udi_hid.c 
         drivers/common/services/usb/class/hid/device/u2f/udi_hid_u2f.c 
-        drivers/common/services/usb/class/hid/device/generic/udi_hid_generic.c 
+        drivers/common/services/usb/class/hid/device/hww/udi_hid_hww.c 
         drivers/common/services/usb/class/composite/device/udi_composite_desc.c 
         drivers/common/services/usb/udc/udc.c 
         drivers/common/utils/interrupt/interrupt_sam_nvic.c 
@@ -140,7 +140,7 @@ set(DRIVER-INCLUDES
         drivers/common/services/usb/class/hid
         drivers/common/services/usb/class/hid/device
         drivers/common/services/usb/class/hid/device/u2f
-        drivers/common/services/usb/class/hid/device/generic
+        drivers/common/services/usb/class/hid/device/hww
         drivers/common/services/usb/class/composite/device
         drivers/common/services/usb/udc
         drivers/common/utils
